Move modal to side, clean up look, move state into modal
mix of vintage coding and intellij ai for the ui changes.
