--- components/modal.tsx
@@ -1,62 +1,0 @@
-import { ReactNode } from 'react';
-
-// <button
-//   className="rounded-md bg-gray-950/5 px-2.5 py-1.5 text-sm font-semibold text-gray-900 hover:bg-gray-950/10">Open
-//   dialog
-// </button>
-type ModalProps = {
-  isOpen: boolean;
-  closeModal: () => void;
-  onSubmit: () => void;
-  submitText: string;
-  children: ReactNode;
-};
-export default function Modal({
-  isOpen,
-  closeModal,
-  submitText,
-  onSubmit,
-  children,
-}: ModalProps) {
-  return !isOpen ? (
-    <></>
-  ) : (
-    <div>
-      <div
-        className="relative z-10"
-        aria-labelledby="dialog-title"
-        role="dialog"
-        aria-modal="true"
-      >
-        <div
-          className="fixed inset-0 bg-gray-500/75 transition-opacity"
-          aria-hidden="true"
-        ></div>
-
-        <div className="fixed inset-0 z-10 w-screen overflow-y-auto">
-          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-            <div className="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
-              {children}
-              <div className="bg-gray-50 px-4 py-3 sm:flex sm:flex-row-reverse sm:px-6">
-                <button
-                  type="button"
-                  onClick={onSubmit}
-                  className="inline-flex w-full justify-center rounded-md bg-red-600 px-3 py-2 text-sm font-semibold text-white shadow-xs hover:bg-red-500 sm:ml-3 sm:w-auto"
-                >
-                  {submitText}
-                </button>
-                <button
-                  type="button"
-                  onClick={closeModal}
-                  className="mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-xs ring-1 ring-gray-300 ring-inset hover:bg-gray-50 sm:mt-0 sm:w-auto"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
